+++ src/index.ts
@@ -2,11 +2,7 @@ export * as api from "./api";
 export * from "./api";
 
 export { default as Playthrough, ConnectionStatus } from "./Playthrough";
-export {
-  default as Microphone,
-  SpeechRecognitionOptions,
-  SpeechRecognitionErrorCode,
-} from "./Microphone";
+export { default as Microphone, SpeechRecognitionOptions } from "./Microphone";
 export { default as Speaker } from "./Speaker";
 
 export {